+++ lib/inference.ml
@@ -207,11 +207,12 @@ let search (oracle: string option -> Oracle.oracle_status) (unclean_g: grammar) 
 
     let start_time = Unix.gettimeofday () in
     Oracle.call_time := 0.;
+    Oracle.idle_time := 0.;
     Quotient.call_time := 0.;
 
     let print_end_time () =
         let total_duration = Unix.gettimeofday () -. start_time in
-        Log.L.info (fun m -> m "Search duration: %.2fs (inference: %.2fs, quotient: %.2fs, oracle: %.2fs)." total_duration (total_duration -. !Quotient.call_time -. !Oracle.call_time) (!Quotient.call_time) !Oracle.call_time);
+        Log.L.info (fun m -> m "Search duration: %.2fs (inference: %.2fs, quotient: %.2fs, oracle: %.2fs, idle: %.2fs)." total_duration (total_duration -. !Quotient.call_time -. !Oracle.call_time -. !Oracle.idle_time) !Quotient.call_time !Oracle.call_time !Oracle.idle_time);
         Log.L.info (fun m -> m "%d calls to oracle." !Oracle.call_nb) in
 
     let finalize () =
@@ -220,8 +221,8 @@ let search (oracle: string option -> Oracle.oracle_status) (unclean_g: grammar) 
         Option.iter (fun ch -> Log.L.info (fun m -> m "Save search graph."); output_string ch "}"; close_out ch) graph_channel;
         Option.iter (fun ch -> Log.L.info (fun m -> m "Save quotient graph."); output_string ch "}"; close_out ch) qgraph_channel;
         match h_fname with
-        | Some fname when Hashtbl.length heuristic > initial_heuristic_length -> Log.L.info (fun m -> m "Save heuristic values into %s" fname); Marshal.to_channel (open_out_bin fname) heuristic []
-        | _ -> () in
+            | Some fname when Hashtbl.length heuristic > initial_heuristic_length -> Log.L.info (fun m -> m "Save heuristic values into %s" fname); Marshal.to_channel (open_out_bin fname) heuristic []
+            | _ -> () in
 
     let inj = start in