+++ lib/oracle.ml
@@ -8,7 +8,7 @@ open Grammar
  *)
 type oracle_status = Syntax_error | (*Semantic_error |*) No_error | Grammar_error
 
-let call_time = ref 0. and call_nb = ref 0
+let call_time = ref 0. and idle_time = ref 0. and call_nb = ref 0 and last_oracle_call = ref 0.
 
 let string_of_oracle_status (s: oracle_status) : string = match s with
     | Syntax_error -> "Syntax error"
@@ -31,11 +31,8 @@ let oracle_mem_no_option (o: string -> oracle_status) : string -> oracle_status 
             Log.L.debug (fun m -> m "Oracle memoization: %s" inj);
             Hashtbl.find mem inj
         end else begin
-            let start_time = Unix.gettimeofday () in
-            call_nb := !call_nb + 1;
             let answer = o inj in
             Hashtbl.add mem inj answer;
-            call_time := !call_time +. (Unix.gettimeofday () -. start_time);
             (*if verbose then print_endline ((string_of_int (Hashtbl.length mem))^"th call to oracle: "^inj^" ("^(string_of_oracle_status answer)^")");*)
             answer
         end
@@ -46,19 +43,35 @@ let handle_option (oracle : string -> oracle_status): string option -> oracle_st
     | Some inj -> oracle inj
 
 (* construct an oracle from a script *)
-let oracle_from_script (timeout: float option) (fname: string) (inj: string) : oracle_status =
+let oracle_from_script (interval: float option) (timeout: float option) (fname: string) (inj: string) : oracle_status =
     let escape_char (c: char) : string = match c with
         | '\'' -> "'\"'\"'" (* in single quote string, only ' must be escaped *)
         | _ -> String.make 1 c in
     let esc_inj = (string_of_list "" "" escape_char (List.init (String.length inj) (String.get inj))) in
+
     let prefix = match timeout with
         | Some n -> "timeout "^(string_of_float n)^" "
         | None -> "" in
+
+    if interval <> None then begin
+        let sleep = Option.get interval -. (Unix.gettimeofday () -. !last_oracle_call) in
+        last_oracle_call := Unix.gettimeofday ();
+        if sleep > 0. then begin
+            Log.L.debug (fun m -> m "Wait %.2fs before oracle call" sleep);
+            idle_time := !idle_time +. sleep;
+            Unix.sleepf sleep
+        end
+    end;
+
     let cmd = match Logs.Src.level Log.poirotsrc with
         | Some Logs.Debug -> fname^" \'"^esc_inj^"\'"
         | _ -> fname^" \'"^esc_inj^"\' >/dev/null 2>&1" in
+
     Log.L.debug (fun m -> m "Call to oracle: %s." cmd);
+    call_nb := !call_nb + 1;
+    let start_time = Unix.gettimeofday () in
     let error_code = Sys.command (prefix^cmd) in
+    call_time := !call_time +. (Unix.gettimeofday () -. start_time);
     let answer = oracle_status_of_int error_code in
     Log.L.info (fun m -> m "Oracle answer to %s: %s" inj (string_of_oracle_status answer));
     answer
@@ -67,4 +80,4 @@ let oracle_from_script (timeout: float option) (fname: string) (inj: string) : o
 let oracle_mem (oracle: string -> oracle_status) : (string option -> oracle_status) = handle_option (oracle_mem_no_option oracle)
 
 (* construct an oracle from a script *)
-let oracle_mem_from_script (timeout: float option) (fname: string) : (string option -> oracle_status) = oracle_mem (oracle_from_script timeout fname)
+let oracle_mem_from_script (interval: float option) (timeout: float option) (fname: string) : (string option -> oracle_status) = oracle_mem (oracle_from_script interval timeout fname)