+++ lib/poirot.mli
@@ -37,7 +37,7 @@ val quotient : ?oneline_comment: string option -> ?qgraph_fname: string option -
 (** {1 Oracle functions} *)
 
 (** [make_oracle_from_script filename] returns an oracle based on the script [filename]. The script must return error code 0 if the injection is lexically correct, 1 otherwise. *)
-val make_oracle_from_script : ?timeout: float option -> string -> (string option -> oracle_status)
+val make_oracle_from_script : ?interval: float option -> ?timeout: float option -> string -> (string option -> oracle_status)
 
 (** [make_oracle_from_fun f] returns an oracle based on the function [f]. The function must return 0 if the injection is lexically correct, 1 otherwise. *)
 val make_oracle_from_fun : (string -> int) -> (string option -> oracle_status)