+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - ANTLR4 export more stable
 - Dune installs `quotient_poirot`
 - Heuristic save can be disabled
+- Can set a minimal interval between two oracle calls
 
 # v0.3
 - Dune installs `poirot` and `bnf2antlr4`